# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/temples/{templeId}/gallery/presigned-url": {
      "post": {
        "operationId": "GalleryController_getPresignedUploadUrl",
        "summary": "Generate pre-signed S3/Cloudinary upload signature/URL (staff+)",
        "parameters": [
          {
            "name": "templeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Gallery"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/temples/{templeId}/gallery": {
      "get": {
        "operationId": "GalleryController_findByTemple",
        "summary": "List gallery items for a temple (public)",
        "parameters": [
          {
            "name": "templeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isPublished",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gallery"
        ]
      },
      "post": {
        "operationId": "GalleryController_create",
        "summary": "Create gallery item (staff+)",
        "parameters": [
          {
            "name": "templeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Gallery"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/temples/{templeId}/gallery/{id}": {
      "get": {
        "operationId": "GalleryController_findById",
        "summary": "Get gallery item by ID (public)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gallery"
        ]
      },
      "put": {
        "operationId": "GalleryController_update",
        "summary": "Update gallery item (staff+)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gallery"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "GalleryController_delete",
        "summary": "Delete gallery item (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gallery"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/gurukul": {
      "get": {
        "operationId": "GurukulController_getGurukul",
        "summary": "Get Gurukul identity, overview and daily schedule (public)",
        "parameters": [
          {
            "name": "templeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gurukul"
        ]
      }
    },
...
    "/api/v1/gurukul/dincharya": {
      "get": {
        "operationId": "GurukulController_getDincharya",
        "summary": "Get Gurukul Dincharya (daily routine schedule) (public)",
        "parameters": [
          {
            "name": "gurukulId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gurukul"
        ]
      }
    },
    "/api/v1/gurukul/admissions": {
      "post": {
        "operationId": "GurukulController_createAdmission",
        "summary": "Submit Gurukul admission / Pravesh application (public)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Gurukul"
        ]
      }
    },
    "/api/v1/admin/gurukul": {
      "get": {
        "operationId": "AdminGurukulController_adminGetGurukul",
        "summary": "Get Gurukul details with schedules & counts (admin)",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gurukul Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/gurukul/{id}": {
      "put": {
        "operationId": "AdminGurukulController_adminUpdateGurukul",
        "summary": "Update Gurukul overview and guidelines (admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGurukulDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gurukul Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/gurukul/admissions": {
      "get": {
        "operationId": "AdminGurukulController_adminGetAdmissions",
        "summary": "List admission applications with filters (admin)",
        "parameters": [
          {
            "name": "gurukulId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gurukul Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/gurukul/admissions/{id}": {
      "get": {
        "operationId": "AdminGurukulController_adminGetAdmissionById",
        "summary": "Get single admission application details (admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gurukul Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "AdminGurukulController_adminUpdateAdmission",
        "summary": "Review and update admission status & notes (admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdmissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gurukul Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/gurukul/schedule": {
      "post": {
        "operationId": "AdminGurukulController_adminCreateSchedule",
        "summary": "Create new Dincharya schedule entry (admin)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Gurukul Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/gurukul/schedule/{id}": {
      "put": {
        "operationId": "AdminGurukulController_adminUpdateSchedule",
        "summary": "Update Dincharya schedule entry (admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gurukul Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminGurukulController_adminDeleteSchedule",
        "summary": "Delete Dincharya schedule entry (admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Gurukul Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Temple Digital Platform API",
    "description": "Comprehensive RESTful Backend API for Temple Digital Platform (Devotee & Admin Panel)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateAdmissionDto": {
        "type": "object",
        "properties": {
          "gurukulId": {
            "type": "string"
          },
          "studentName": {
            "type": "string",
            "example": "Aarav Sharma"
          },
          "guardianName": {
            "type": "string",
            "example": "Ramesh Sharma"
          },
          "phone": {
            "type": "string",
            "example": "+919876543210"
          },
          "email": {
            "type": "string",
            "example": "parent@example.com"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "2012-05-15"
          },
          "previousEducation": {
            "type": "string",
            "example": "Class 5 passed"
          },
          "address": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "example": "Seeking admission for Vedic studies"
          }
        },
        "required": [
          "studentName",
          "guardianName",
          "phone"
        ]
      },
      "UpdateGurukulDto": {
        "type": "object",
        "properties": {
          "templeId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Shree Neelkantheshwar Mahadev Ved Vedang Gurukulam"
          },
          "description": {
            "type": "string"
          },
          "about": {
            "type": "string"
          },
          "philosophy": {
            "type": "string"
          },
          "admissionInfo": {
            "type": "string"
          },
          "contactInfo": {
            "type": "string"
          },
          "rules": {
            "type": "string"
          },
          "isPublished": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "UpdateAdmissionDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "REVIEWING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "adminNotes": {
            "type": "string"
          }
        }
      },
      "CreateScheduleDto": {
        "type": "object",
        "properties": {
          "gurukulId": {
            "type": "string"
          },
          "activityName": {
            "type": "string",
            "example": "Pratah Smaran"
          },
          "description": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "example": "04:00 AM"
          },
          "endTime": {
            "type": "string",
            "example": "05:00 AM"
          },
          "displayOrder": {
            "type": "number",
            "default": 0
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "gurukulId",
          "activityName",
          "startTime"
        ]
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {
          "gurukulId": {
            "type": "string"
          },
          "activityName": {
            "type": "string",
            "example": "Pratah Smaran"
          },
          "description": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "example": "04:00 AM"
          },
          "endTime": {
            "type": "string",
            "example": "05:00 AM"
          },
          "displayOrder": {
            "type": "number",
            "default": 0
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
    }
  }
}